Decompile Fishhead `func_us_801D8DF0` + `func_us_801D8FE0`
