+++ config/splat.us.stno4.yaml
@@ -65,7 +65,8 @@ segments:
       - [0x40BF0, .rodata, e_stage_name]
       - [0x40C24, .rodata, e_toad]
       - [0x40C40, .rodata, e_frog]
-      - [0x40C6C, .rodata, unk_55678]
+      - [0x40C6C, .rodata, e_fishhead]
+      - [0x40C84, .rodata, unk_59560]
       - [0x40CB8, .rodata, e_spear_guard]
       - [0x40CEC, .rodata, e_bone_archer]
       - [0x40CF8, .rodata, e_skeleton_ape]
@@ -88,7 +89,8 @@ segments:
       - [0x55678, c, water_effects]
       - [0x5798C, c, e_toad]
       - [0x57FAC, c, e_frog]
-      - [0x585A4, c, unk_55678]
+      - [0x585A4, c, e_fishhead]
+      - [0x59560, c, unk_59560]
       - [0x5BCDC, c, e_spear_guard]
       - [0x5CA9C, c, e_bone_archer]
       - [0x5D688, c, e_skeleton_ape]
