+++ include/entity.h
@@ -2938,7 +2938,7 @@ typedef struct {
 } ET_FrogToad;
 
 typedef struct {
-    /* 0x7C */ s32 : 32;
+    /* 0x7C */ struct Primitive* prim;
     /* 0x80 */ s16 unk80;
     /* 0x82 */ s16 unk82;
     /* 0x84 */ u8 unk84;
