+++ src/st/no4/no4.h
@@ -67,9 +67,9 @@ typedef enum EntityIDs {
     /* 0x3A */ E_UNK_NO4_SPEAR_GUARD = 0x3A,
     /* 0x3B */ E_SPEAR_GUARD_BLOCK,
     /* 0x3C */ E_TOAD,
-    // /* 0x3D */ E_ID_3D,
-    /* 0x3E */ E_FISHHEAD = 0x3E,
-    // /* 0x3F */ E_ID_3F,
+    /* 0x3D */ E_FISHHEAD_SPAWNER,
+    /* 0x3E */ E_FISHHEAD,
+    /* 0x3F */ E_FISHHEAD_FIREBALL,
     // /* 0x40 */ E_ID_40,
     // /* 0x41 */ E_ID_41,
     // /* 0x42 */ E_ID_42,
